Fix for premake and including files in sandbox from game
