+++ StortSpelProjekt/premake5.lua
@@ -82,7 +82,7 @@ project "Sandbox"
     language "C++"
     targetdir "bin/%{cfg.buildcfg}/%{prj.name}"
     objdir "bin-int/%{cfg.buildcfg}/%{prj.name}"
-    files { "%{prj.location}/src/**.cpp", "%{prj.location}/src/**.h", "%{prj.location}/src/**.hlsl" }
+    files { "%{prj.location}/src/**.cpp", "%{prj.location}/src/**.h", "%{prj.location}/src/**.hlsl", "%{prj.location}/../Game/src/files/**.cpp", "%{prj.location}/../Game/src/files/**.h" }
     
     filter { "files:**.hlsl" }
         flags "ExcludeFromBuild"
@@ -90,7 +90,7 @@ project "Sandbox"
     filter "configurations:*"
         cppdialect "C++17"
     
-    includedirs {"Vendor/Include/", "Engine/src/", "Engine/src/Headers/"}
+    includedirs {"Vendor/Include/", "Engine/src/", "Engine/src/Headers/", "Game/src/files/" }
     libdirs { "Vendor/Lib/**" }
     links {
         "Engine"
